rotation_3d was supposed to be in pyvgl_algo, not pyvgl.
