+++ pyvpgl.cxx
@@ -115,7 +115,7 @@ void wrap_vpgl(py::module &m)
 {
   py::class_<vpgl_proj_camera<double> >(m, "proj_camera")
     .def(py::init<vnl_matrix_fixed<double,3,4> >())
-    .def("__str__", stream2str<vpgl_proj_camera<double> >)
+    .def("__str__", streamToString<vpgl_proj_camera<double> >)
     .def("project", vpgl_project_homg_point<vpgl_proj_camera<double> >)
     .def("project", vpgl_project_point<vpgl_proj_camera<double> >)
     .def("project", vpgl_project_vector<vpgl_proj_camera<double> >)
@@ -144,7 +144,7 @@ void wrap_vpgl(py::module &m)
 
   py::class_<vpgl_perspective_camera<double>, vpgl_proj_camera<double> >(m, "perspective_camera")
     .def(py::init<vpgl_calibration_matrix<double>, vgl_rotation_3d<double>, vgl_vector_3d<double> >())
-    .def("__str__", stream2str<vpgl_perspective_camera<double> >);
+    .def("__str__", streamToString<vpgl_perspective_camera<double> >);
 
   py::class_<vpgl_scale_offset<double> >(m, "scale_offset")
     .def(py::init<>())
@@ -175,7 +175,7 @@ void wrap_vpgl(py::module &m)
                    double, double, double, double, double, double, double, double, double, double>())
      .def(py::init<vnl_matrix_fixed<double,4,20>, std::vector<vpgl_scale_offset<double> > >())
      .def("copy", &vpgl_rational_camera<double>::clone)
-     .def("__str__", stream2str<vpgl_rational_camera<double> >)
+     .def("__str__", streamToString<vpgl_rational_camera<double> >)
      .def("save", &vpgl_rational_camera<double>::save)
      .def("coefficient_matrix", &vpgl_rational_camera<double>::coefficient_matrix)
      .def("scale_offsets", &vpgl_rational_camera<double>::scale_offsets)
@@ -262,7 +262,7 @@ void wrap_vpgl(py::module &m)
         py::arg("cs_name")=vpgl_lvcs::wgs84,py::arg("ang_unit")=vpgl_lvcs::DEG,py::arg("elev_unit")=vpgl_lvcs::METERS)
 
     // python print
-    .def("__str__", stream2str<vpgl_lvcs>)
+    .def("__str__", streamToString<vpgl_lvcs>)
 
     // getters
     .def("get_origin",     [](vpgl_lvcs &L) {double lon,lat,e; L.get_origin(lat,lon,e); return std::make_tuple(lon,lat,e); })
@@ -371,7 +371,7 @@ void wrap_vpgl(py::module &m)
   py::class_<vpgl_geo_camera>(m, "geo_camera")
     // Default methods
     .def(py::init<>())
-    .def("__str__", stream2str<vpgl_geo_camera >)
+    .def("__str__", streamToString<vpgl_geo_camera >)
     // Convert pixel coords (u,v) to a lon/lat pair
     .def("img_to_global",
       [](vpgl_geo_camera &G, double const u, double const v)
